requirements.txt
langchain
wikipedia
langchain-community
langchain-google-genai
python-dotenv
pydantic
duckduckgo-search